refactor: Добавлены интерфейсы на сервисном слое style: убраны излишние Autowired style: убраны группировки пакетов при импорте style: доблен слой работы с данными
