+++ src/main/java/ru/yandex/practicum/filmorate/service/UserService.java
@@ -1,83 +1,24 @@
 package ru.yandex.practicum.filmorate.service;
 
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
 import ru.yandex.practicum.filmorate.model.User;
-import ru.yandex.practicum.filmorate.storage.UserStorage;
 
-import java.util.ArrayList;
 import java.util.List;
-import java.util.Set;
-import java.util.stream.Collectors;
 
-@Service
-public class UserService {
-    private final UserStorage userStorage;
+public interface UserService {
 
-    @Autowired
-    public UserService(UserStorage userStorage) {
-        this.userStorage = userStorage;
-    }
+    List<User> getUsers();
 
-    public List<User> getUsers() {
-        return userStorage.getUsers();
-    }
+    User getUser(Long id);
 
-    public User getUser(Long id) {
-        return userStorage.getUser(id);
-    }
+    User addUser(User user);
 
-    public User addUser(User user) {
-        return userStorage.addUser(user);
-    }
+    User updateUser(User user);
 
-    public User updateUser(User user) {
-        return userStorage.updateUser(user);
-    }
+    List<User> getFriends(Long id);
 
-    public List<User> getFriends(Long id) {
-        User user = userStorage.getUser(id);
-        return user.getFriends().stream()
-                .map(userStorage::getUser)
-                .collect(Collectors.toList());
-    }
+    List<User> addFriend(Long idUser, Long idFriend);
 
-    public List<User> addFriend(Long idUser, Long idFriend) {
-        User user = userStorage.getUser(idUser);
-        User friend = userStorage.getUser(idFriend);
-        user.addFriend(idFriend);
-        friend.addFriend(idUser);
+    List<User> removeFriend(Long idUser, Long idFriend);
 
-        return user.getFriends().stream()
-                .map(userStorage::getUser)
-                .collect(Collectors.toList());
-    }
-
-    public List<User> removeFriend(Long idUser, Long idFriend) {
-        User user = userStorage.getUser(idUser);
-        User friend = userStorage.getUser(idFriend);
-        user.removeFriend(idFriend);
-        friend.removeFriend(idUser);
-
-        return user.getFriends().stream()
-                .map(userStorage::getUser)
-                .collect(Collectors.toList());
-    }
-
-    public List<User> getCommonFriends(Long idUser, Long idFriend) {
-        User user = userStorage.getUser(idUser);
-        User friend = userStorage.getUser(idFriend);
-
-        if (user.getFriends() == null || friend.getFriends() == null) {
-            return new ArrayList<>();
-        }
-
-        Set<Long> userFriends = user.getFriends();
-        Set<Long> friendFriends = friend.getFriends();
-        userFriends.retainAll(friendFriends);
-
-        return userFriends.stream()
-                .map(userStorage::getUser)
-                .collect(Collectors.toList());
-    }
+    List<User> getCommonFriends(Long idUser, Long idFriend);
 }
